feat(ironfish): Add explicit payout period to mining pool
This adds a "payout period" with a fixed duration, such that any blocks or
shares found during that time will be tied to that period. This will
allow the pool to calculate the exact amount to payout each share in a
deterministic fashion. Materializing explicit rows instead of relying on
timestamps allows the code to be a bit more explicit and easier to
reason about.