feat(ironfish): mining pool tracks status of transactions its created
This introduces a new table which allows the pool to insert a row
whenever it makes a new transaction. The event loop will update the
status of any unconfirmed and unexpired transactions. This allows the
pool to know what transactions it has made, as well as giving it the
ability to re-pay out transactions in the event of one expiring.